pyproject.toml
[tool.poetry]
name = "cogs118a-group051"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "~3.10"
numpy = "^1.22.3"
matplotlib = "^3.5.1"
scikit-learn = "^1.0.2"
scikit-image = "^0.19.2"
tqdm = "^4.64.0"
python-chess = "^1.999"
torch = {version = "^1.11.0+cu113", source = "pytorch"}
pytorch-lightning = "^1.6.4"

[tool.poetry.group.dev.dependencies]
torch-tb-profiler = "^0.4.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.11.0"
black = "^22.3.0"
click = "^8.1.3"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu113"
default = false
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
